Changes for Role Access for Super console and Project delete
git-svn-id: https://opsvcs.tdsops.com/repos/tdstm/trunk@485 fecba46a-3362-0410-80e5-8c76cdb905a0

+++ grails-app/controllers/ProjectController.groovy
@@ -25,7 +25,10 @@ class ProjectController {
         } else { 
         	def currProj = session.getAttribute("CURR_PROJ");
         	def currProjectInstance = Project.get( currProj.CURR_PROJ )
-        	def projectLogo = ProjectLogo.findByProject(currProjectInstance)
+        	def projectLogo
+        	if(currProjectInstance){
+        		projectLogo = ProjectLogo.findByProject(currProjectInstance)
+        	}
         	def imageId
         	if(projectLogo){
         		imageId = projectLogo.id
@@ -51,14 +54,21 @@ class ProjectController {
     }
 
     def delete = {
-        def projectInstance = Project.get( params.id )
-        if(projectInstance) {
-            projectInstance.delete()
-            flash.message = "Project ${projectInstance} deleted"
-            redirect(action:list)
-        }
-        else {
-            flash.message = "Project not found with id ${params.id}"
+    	
+    	def currProj = session.getAttribute("CURR_PROJ").CURR_PROJ;
+        if(currProj != params.id){
+        	def projectInstance = Project.get( params.id )
+	        if(projectInstance) {
+	            projectInstance.delete()
+	            flash.message = "Project ${projectInstance} deleted"
+	            redirect(action:list)
+	        }
+	        else {
+	            flash.message = "Project not found with id ${params.id}"
+	            redirect(action:list)
+	        }
+        } else {
+            flash.message = "Unable to Delete the Current Project"
             redirect(action:list)
         }
     }
